Begin work on the parser.
Initial commit for the project. Begin work on the parser as described in Write
Yourself a Scheme [1]. Many problems are present. For example, only a few of
the types of numbers described in the tutorial are handled and even then they
are sometimes handled erroneously (like floating point numbers needing a
leading number, or a decimal point even if the number is given as inexact).

References:
[1] http://en.wikibooks.org/wiki/Write_Yourself_a_Scheme_in_48_Hours/Parsing